Remove event fired for previous files with multiple:false. Close #3878

+++ src/kendo.upload.js
@@ -342,7 +342,9 @@ var __meta__ = {
             }
 
             if (!that.multiple && existingFileEntries.length > 0) {
-                that._module.onRemove({target : $(existingFileEntries, that.wrapper)});
+                if (!that.trigger(REMOVE, { files: existingFileEntries.data("fileNames") })) {
+                    that._module.onRemove({target : $(existingFileEntries, that.wrapper)});
+                }
             }
 
             return fileEntry;
